backend/requirements.txt
Django==5.1.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-redis==5.4.0
PyMySQL>=1.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
Pillow>=10.1.0
opencv-python>=4.8.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0
